Move ui to view package
