+++ src/main/java/easton/Easton.java
@@ -4,6 +4,7 @@ import easton.model.Deadline;
 import easton.model.Event;
 import easton.model.Task;
 import easton.model.ToDo;
+import easton.view.Ui;
 
 import java.io.*;
 import java.util.ArrayList;
